Work around change to Bjam behavior.
As of version 1.55, Boost Build stopped allowing a Python script to be
specified as the path to the compiler (possibly due to a change in which
Windows API they called, judging from the error). This commit adds a C++
program which calls Python with it's argument list after modifying it's
zeroth argument to end in ".py". It's a crude fix but it seems to work
just fine.